+++ packages/amplify-graphql-api-construct-tests/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/aws-amplify/amplify-category-api/compare/amplify-graphql-api-construct-tests@1.2.0...amplify-graphql-api-construct-tests@1.2.1) (2023-12-14)
+
+**Note:** Version bump only for package amplify-graphql-api-construct-tests
+
 # [1.2.0](https://github.com/aws-amplify/amplify-category-api/compare/amplify-graphql-api-construct-tests@1.1.3...amplify-graphql-api-construct-tests@1.2.0) (2023-12-06)
 
 ### Features