+++ packages/amplify-graphql-api-construct/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-api-construct@1.5.0...@aws-amplify/graphql-api-construct@1.5.1) (2023-12-14)
+
+**Note:** Version bump only for package @aws-amplify/graphql-api-construct
+
 # [1.5.0](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-api-construct@1.4.3...@aws-amplify/graphql-api-construct@1.5.0) (2023-12-06)
 
 ### Bug Fixes