+++ packages/amplify-graphql-transformer-migrator/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.14](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-migrator@2.2.13...@aws-amplify/graphql-transformer-migrator@2.2.14) (2023-12-14)
+
+**Note:** Version bump only for package @aws-amplify/graphql-transformer-migrator
+
 ## [2.2.13](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-migrator@2.2.12...@aws-amplify/graphql-transformer-migrator@2.2.13) (2023-12-06)
 
 **Note:** Version bump only for package @aws-amplify/graphql-transformer-migrator